This vga code can draw a line.
I got the vga_adapter code from Jacob Gardner, but he did not make it. I made the display module and learned how to draw dots and lines. The current code is a failed attempt at drawing a rectangle. The included jpg is a picture of the current output of the display module.