Module that creates a simple rectangle.
Can be instantiated in display.v instead of replacing display.v. Uses an FSM to draw pixels of certain colors onto the screen in desired locations. The FSM essentially implements a nested for loop where the x pixels iterate in the inner loop and the y pixels iterate in the outer loop. Might potentially have some timing issues later on due to the number of states.