Add files via upload
There is no VGA output for this file and I don't know why. It is not because I changed the vga.v module, because it did run with that change applied. It stopped outputting when I changed some && to || just to see what would happen. And then it kept not outputting anything after I undid that change, so I don't know what's wrong.