The red block continuously moves right by 5 pixels
This is what I was trying to accomplish however I have no idea why this is the output of this code. The x-coordinates of the block should only change in the INIT state and the FSM should never go back to the INIT state after leaving. There is also no code for what to do after the DRAW2 state ends so I don't know what is happening there.